\name{dots}
\alias{dots}

\title{One dimensional dots dataset --- outlier detection example}

\usage{data(dots)}

\description{
A matrix whose columns are the \code{Z} and \code{W} \code{dots} datasets from
Wilkinson (2016).
}
\section{References}{
  L. Wilkinson. 2016. Vizualizing Outliers. 
<https://www.cs.uic.edu/~wilkinson/Publications/outliers.pdf>.
}

\keyword{datasets}